requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
web3>=6
eth-account
pydantic>=2
python-telegram-bot[webhooks]
